LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/xhr-in-sharedworker-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'message_port.start')

Harness Error (FAIL), message = TypeError: null is not an object (evaluating 'message_port.start')
